refactor: hooks 디렉토리 언더바 처리
